GPF-1606 flexible cnv loader reorganization
