+++ dae/dae/backends/cnv/loader.py
@@ -75,16 +75,13 @@ import logging
 import argparse
 from pathlib import Path
 
-from typing import Callable, List, Optional, Dict, Any, Tuple, Generator, \
+from typing import List, Optional, Dict, Any, Tuple, Generator, \
     Union, TextIO, cast
 
-from copy import copy, deepcopy
-from dae.backends.raw.flexible_variant_loader import adjust_chrom_prefix, \
-    flexible_variant_loader
-import numpy as np
+from copy import copy
+from dae.backends.cnv.flexible_cnv_loader import flexible_cnv_loader
 import pandas as pd
 
-from dae.variants.core import Allele
 from dae.genomic_resources.reference_genome import ReferenceGenome
 from dae.backends.raw.loader import VariantsGenotypesLoader, TransmissionType
 
@@ -95,329 +92,9 @@ from dae.variants.family_variant import FamilyVariant
 from dae.backends.raw.loader import CLIArgument
 
 from dae.utils.regions import Region
-from dae.utils.variant_utils import get_interval_locus_ploidy
 
 
 logger = logging.getLogger(__name__)
-
-
-def _cnv_location_to_vcf_trasformer() \
-        -> Callable[[Dict[str, Any]], Dict[str, Any]]:
-    """
-    In case the input uses CNV location this transormer will product
-    internal (chrom, pos, pos_end) description of the CNV position.
-    """
-    def trasformer(result: Dict[str, Any]) -> Dict[str, Any]:
-        location = result["location"]
-        chrom, range = location.split(":")
-        beg, end = range.split("-")
-        result["chrom"] = chrom
-        result["pos"] = int(beg)
-        result["pos_end"] = int(end)
-
-        return result
-
-    return trasformer
-
-
-def _cnv_vcf_to_vcf_trasformer() \
-        -> Callable[[Dict[str, Any]], Dict[str, Any]]:
-    """
-    In case the input uses VCF-like description of the CNVs this
-    transformer will chec it and handle the proper type conversion for
-    `pos` and `pos_end` values.
-    """
-    def trasformer(result: Dict[str, Any]) -> Dict[str, Any]:
-        chrom = result["chrom"]
-        pos = int(result["pos"])
-        pos_end = int(result["pos_end"])
-
-        result["chrom"] = chrom
-        result["pos"] = pos
-        result["pos_end"] = pos_end
-
-        return result
-
-    return trasformer
-
-
-def _configure_cnv_location(
-        header: List[str],
-        transformers: List[Callable[[Dict[str, Any]], Dict[str, Any]]],
-        cnv_chrom: Optional[str] = None,
-        cnv_start: Optional[str] = None,
-        cnv_end: Optional[str] = None,
-        cnv_location: Optional[str] = None) -> None:
-    """
-    This helper function will **configure** the header and transformers needed
-    to handle position of CNVs in the input record.
-    """
-    if cnv_chrom is not None or cnv_start is not None or \
-            cnv_end is not None:
-        if cnv_location is not None:
-            raise ValueError(
-                f"mixed variant location definitions: "
-                f"vcf({cnv_chrom}:{cnv_start}-{cnv_end}) and "
-                f"location({cnv_location})")
-        if cnv_chrom is None:
-            cnv_chrom = "chrom"
-        if cnv_start is None:
-            cnv_start = "pos"
-        if cnv_end is None:
-            cnv_end = "pos_end"
-
-        chrom_index = header.index(cnv_chrom)
-        start_index = header.index(cnv_start)
-        end_index = header.index(cnv_end)
-
-        if chrom_index == -1 or start_index == -1 or end_index == -1:
-            raise ValueError(
-                f"cant find cnv vcf position like columns "
-                f"vcf({cnv_chrom}:{cnv_start}-{cnv_end}) in header: "
-                f"{header}"
-            )
-        header[chrom_index] = "chrom"
-        header[start_index] = "pos"
-        header[end_index] = "pos_end"
-        transformers.append(_cnv_vcf_to_vcf_trasformer())
-    else:
-        if cnv_location is None:
-            cnv_location = "location"
-        location_index = header.index(cnv_location)
-        if location_index == -1:
-            raise ValueError(
-                f"can find cnv location column "
-                f"location({cnv_location}) in header: "
-                f"{header}"
-            )
-        header[location_index] = "location"
-        transformers.append(_cnv_location_to_vcf_trasformer())
-
-
-def _cnv_dae_best_state_to_best_state(
-        families: FamiliesData, genome: ReferenceGenome) \
-        -> Callable[[Dict[str, Any]], Dict[str, Any]]:
-    """
-    In case the genotype of the CNVs is specified in old
-    dae family id/best state notation, this transformer will handle it
-    and transform it to canonical family id/best state form
-    """
-    def transformer(result: Dict[str, Any]) -> Dict[str, Any]:
-        variant_type = result["variant_type"]
-        actual_ploidy = np.fromstring(
-            result["best_state"], dtype=np.int8, sep=" ")
-        family_id = result["family_id"]
-        family = families[family_id]
-        chrom = result["chrom"]
-        pos = result["pos"]
-        pos_end = result["pos_end"]
-
-        expected_ploidy = np.asarray([
-            get_interval_locus_ploidy(
-                chrom, pos, pos_end, p.sex, genome
-            ) for p in family.members_in_order
-        ])
-        if variant_type == Allele.Type.large_duplication:
-            alt_row = actual_ploidy - expected_ploidy
-        elif variant_type == Allele.Type.large_deletion:
-            alt_row = expected_ploidy - actual_ploidy
-        else:
-            raise ValueError(
-                f"unexpected variant type: {variant_type}")
-        ref_row = expected_ploidy - alt_row
-        best_state = np.stack((ref_row, alt_row)).astype(np.int8)
-        result["best_state"] = best_state
-
-        return result
-
-    return transformer
-
-
-def _cnv_person_id_to_best_state(
-        families: FamiliesData, genome: ReferenceGenome) \
-        -> Callable[[Dict[str, Any]], Dict[str, Any]]:
-
-    """
-    In case the genotype is specified by person id having the variant
-    this transformer will transform it into canonical family id/best state
-    form
-    """
-    def transformer(result: Dict[str, Any]) -> Dict[str, Any]:
-        person_id = result["person_id"]
-        person = families.persons[person_id]
-        family = families[person.family_id]
-
-        chrom = result["chrom"]
-        pos = result["pos"]
-        pos_end = result["pos_end"]
-
-        expected_ploidy = np.asarray([
-            get_interval_locus_ploidy(
-                chrom, pos, pos_end, p.sex, genome
-            ) for p in family.members_in_order
-        ])
-        alt_row = np.zeros(len(family.members_in_order), dtype=np.int8)
-        alt_row[person.index] = 1
-
-        ref_row = expected_ploidy - alt_row
-        best_state = np.stack((ref_row, alt_row)).astype(np.int8)
-        result["best_state"] = best_state
-        result["family_id"] = family.family_id
-
-        return result
-
-    return transformer
-
-
-def _configure_cnv_best_state(
-        header: List[str],
-        transformers: List[Callable[[Dict[str, Any]], Dict[str, Any]]],
-        families: FamiliesData,
-        genome: ReferenceGenome,
-        cnv_person_id: Optional[str] = None,
-        cnv_family_id: Optional[str] = None,
-        cnv_best_state: Optional[str] = None) -> None:
-    """
-    This helper function will **configure** the header and transformers
-    needed to handle family genotypes of CNVs.
-    """
-
-    if cnv_person_id is not None:
-        if cnv_family_id is not None or cnv_best_state is not None:
-            raise ValueError(
-                f"mixed configuration of cnv best state: "
-                f"person_id({cnv_person_id}) <-> "
-                f"family_id({cnv_family_id}) and "
-                f"best_state({cnv_best_state})"
-            )
-        person_index = header.index(cnv_person_id)
-        if person_index == -1:
-            raise ValueError(
-                f"cant find person_id({cnv_person_id}) in header: "
-                f"{header}")
-
-        header[person_index] = "person_id"
-
-        transformers.append(
-            _cnv_person_id_to_best_state(families, genome)
-        )
-    else:
-        if cnv_family_id is None:
-            cnv_family_id = "family_id"
-        if cnv_best_state is None:
-            cnv_best_state = "best_state"
-        family_index = header.index(cnv_family_id)
-        best_state_index = header.index(cnv_best_state)
-        if family_index == -1 or best_state_index == -1:
-            raise ValueError(
-                f"cant find family_id({cnv_family_id}) or "
-                f"best_state({cnv_best_state}) in header: "
-                f"{header}")
-        header[family_index] = "family_id"
-        header[best_state_index] = "best_state"
-
-        transformers.append(
-            _cnv_dae_best_state_to_best_state(families, genome))
-
-
-def _cnv_variant_to_variant_type(
-        cnv_plus_values=set(["CNV+"]),
-        cnv_minus_values=set(["CNV-"])) \
-            -> Callable[[Dict[str, Any]], Dict[str, Any]]:
-    """
-    This transformer is used to transform variant type to canonical
-    inernal representation using :class:`Allele.Type.large_duplication` and
-    :class:`Allele.Type.large_deletion`.
-    """
-    def transformer(result: Dict[str, Any]) -> Dict[str, Any]:
-        variant = result["variant"]
-        if variant in cnv_plus_values:
-            variant_type = Allele.Type.large_duplication
-        elif variant in cnv_minus_values:
-            variant_type = Allele.Type.large_deletion
-        else:
-            raise ValueError(f"unexpected CNV variant type: {variant}")
-
-        result["variant_type"] = variant_type
-        return result
-
-    return transformer
-
-
-def _configure_cnv_variant_type(
-        header: List[str],
-        transformers: List[Callable[[Dict[str, Any]], Dict[str, Any]]],
-        cnv_variant_type: Optional[str] = None,
-        cnv_plus_values: Union[str, List[str]] = ["CNV+"],
-        cnv_minus_values: Union[str, List[str]] = ["CNV-"]):
-
-    if isinstance(cnv_plus_values, str):
-        cnv_plus_values = [cnv_plus_values]
-    if isinstance(cnv_minus_values, str):
-        cnv_minus_values = [cnv_minus_values]
-
-    """
-    This helper function **configures** the header and transformer needed
-    to handle CNV variant type in the input data.
-    """
-    if cnv_variant_type is None:
-        cnv_variant_type = "variant"
-
-    variant_type_index = header.index(cnv_variant_type)
-    if variant_type_index == -1:
-        raise ValueError(
-            f"missing variant type column {cnv_variant_type} in header: "
-            f"{header}")
-    header[variant_type_index] = "variant"
-
-    transformers.append(
-        _cnv_variant_to_variant_type(cnv_plus_values, cnv_minus_values)
-    )
-
-
-def _configure_loader(
-        header: List[str],
-        families: FamiliesData,
-        genome: ReferenceGenome,
-        cnv_chrom: Optional[str] = None,
-        cnv_start: Optional[str] = None,
-        cnv_end: Optional[str] = None,
-        cnv_location: Optional[str] = None,
-        cnv_person_id: Optional[str] = None,
-        cnv_family_id: Optional[str] = None,
-        cnv_best_state: Optional[str] = None,
-        cnv_variant_type: Optional[str] = None,
-        cnv_plus_values: List[str] = ["CNV+"],
-        cnv_minus_values: List[str] = ["CNV-"]) \
-        -> Tuple[
-            List[str],
-            List[Callable[[Dict[str, Any]], Dict[str, Any]]]]:
-    """
-    This funciton configures all headers and transformers needed to
-    handle CNVs input data.
-    """
-
-    transformers: List[
-        Callable[[Dict[str, Any]], Dict[str, Any]]] = []
-    header = deepcopy(header)
-
-    _configure_cnv_location(
-        header, transformers,
-        cnv_chrom, cnv_start, cnv_end,
-        cnv_location
-    )
-
-    _configure_cnv_variant_type(
-        header, transformers,
-        cnv_variant_type, cnv_plus_values, cnv_minus_values)
-
-    _configure_cnv_best_state(
-        header, transformers,
-        families, genome,
-        cnv_person_id,
-        cnv_family_id, cnv_best_state)
-
-    return header, transformers
 
 
 def _cnv_loader(
@@ -445,70 +122,49 @@ def _cnv_loader(
     `DataFrame`.
     """
 
-    if isinstance(filepath_or_buffer, str) or \
-            isinstance(filepath_or_buffer, Path):
-        infile = open(filepath_or_buffer, "rt")
-    else:
-        infile = filepath_or_buffer
+    logger.info("unexpected parameters passed to _cnv_loader: %s", kwargs)
 
-    def line_splitter(ln: str) -> List[str]:
-        return ln.strip("\n\r").split(cnv_sep)
+    variant_generator = flexible_cnv_loader(
+        filepath_or_buffer,
+        families,
+        genome,
+        regions=regions,
+        cnv_chrom=cnv_chrom,
+        cnv_start=cnv_start,
+        cnv_end=cnv_end,
+        cnv_location=cnv_location,
+        cnv_person_id=cnv_person_id,
+        cnv_family_id=cnv_family_id,
+        cnv_best_state=cnv_best_state,
+        cnv_variant_type=cnv_variant_type,
+        cnv_plus_values=cnv_plus_values,
+        cnv_minus_values=cnv_minus_values,
+        cnv_sep=cnv_sep,
+        add_chrom_prefix=add_chrom_prefix,
+        del_chrom_prefix=del_chrom_prefix)
 
-    with infile as infile:
-        line = next(infile)
-        header = line_splitter(line)
+    data = []
+    for record in variant_generator:
+        data.append(record)
 
-        header, transformers = _configure_loader(
-            header,
-            families,
-            genome,
-            cnv_chrom,
-            cnv_start,
-            cnv_end,
-            cnv_location,
-            cnv_person_id,
-            cnv_family_id,
-            cnv_best_state,
-            cnv_variant_type,
-            cnv_plus_values,
-            cnv_minus_values)
+    df: pd.DataFarme = cast(
+        pd.DataFrame,
+        pd.DataFrame.from_records(  # type: ignore
+            data, columns=[
+                "chrom", "pos", "pos_end",
+                "variant_type",
+                "family_id", "best_state"
+            ]))
 
-        transformers.append(adjust_chrom_prefix(
-            add_chrom_prefix, del_chrom_prefix))
+    df = df.sort_values(
+        by=["chrom", "pos", "pos_end"])
 
-        def regions_filter(rec):
-            if len(regions) == 0:
-                return True
-            chrom = rec["chrom"]
-            pos = rec["pos"]
-            return any([r.isin(chrom, pos) for r in regions])
-
-        variant_generator = flexible_variant_loader(
-                infile, header, line_splitter, transformers,
-                filters=[regions_filter])
-
-        data = []
-        for record in variant_generator:
-            data.append(record)
-
-        df: pd.DataFarme = cast(
-            pd.DataFrame,
-            pd.DataFrame.from_records(  # type: ignore
-                data, columns=[
-                    "chrom", "pos", "pos_end",
-                    "variant_type",
-                    "family_id", "best_state"
-                ]))
-
-        df = df.sort_values(
-            by=["chrom", "pos", "pos_end"])
-
-        df = df.rename(
-            columns={
-                "pos": "position",
-                "pos_end": "end_position"
-            })
-        return cast(pd.DataFrame, df)
+    df = df.rename(
+        columns={
+            "pos": "position",
+            "pos_end": "end_position"
+        })
+    return cast(pd.DataFrame, df)
 
 
 class CNVLoader(VariantsGenotypesLoader):
